Extract ImageAvatar and InitialLetterAvatar to their own files.
